utleggs-og-kortskjema-automatisering/requirements.txt
pillow==9.3.0
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==0.21.0
reportlab==3.6.12
requests==2.28.1
